Just some testing
no major changes, just performing some testing.

It turns out that the cause for the exponential lag bug is because the
gridDisplay method is running over and over again, and the number of
times it runs on the nth button press corresponds to the (2n)th
fibonacci number (all zero indexing).

Funny how the fibonacci numbers can show up in a bug like that. Anyway,
I still haven’t figured out what’s causing the bug. It actually happens
with all of the buttons on the main screen (and possibly more I don’t
know of).